+++ apps/client/src/pages/my-history/hooks/use-delete-music-mutation.ts
@@ -1,7 +1,7 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { deleteMusicFromSetList } from '@shared/apis/my-history/setlist';
-import { SETLIST_QUERY_KEY } from '@shared/apis/my-history/setlist-queries';
+import { SETLIST_QUERY_KEY } from '@shared/constants/query-key';
 
 export const useDeleteMusicMutation = (setlistId: number) => {
   const queryClient = useQueryClient();
