+++ apps/client/src/pages/my/hooks/use-logout.ts
@@ -2,7 +2,7 @@ import { useNavigate } from 'react-router-dom';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { postLogout } from '@shared/apis/auth/auth';
-import { USER_QUERY_KEY } from '@shared/apis/user/user-queries';
+import { USER_QUERY_KEY } from '@shared/constants/query-key';
 import { routePath } from '@shared/router/path';
 import { BaseResponse } from '@shared/types/api';
 import { authTokenHandler } from '@shared/utils/token-handler';
