+++ apps/client/src/pages/time-table/hooks/use-get-festival-to-add.ts
@@ -1,10 +1,10 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 
-import { GET_FESTIVAL_TO_ADD_QUERY_OPTIONS } from '@shared/apis/confeti-detail/get-festival-to-add-queries';
+import { PERFORMANCE_QUERY_OPTIONS } from '@shared/apis/confeti-detail/performance-queries';
 
 export const useGetFestivalToAdd = () => {
   const { data, fetchNextPage, hasNextPage } = useInfiniteQuery({
-    ...GET_FESTIVAL_TO_ADD_QUERY_OPTIONS.LIST(),
+    ...PERFORMANCE_QUERY_OPTIONS.GET_FESTIVAL_TO_ADD_LIST(),
     getNextPageParam: (lastPage) => {
       return lastPage.nextCursor === -1 ? undefined : lastPage.nextCursor;
     },
