+++ apps/client/src/pages/time-table/hooks/use-timetable-festival-mutation.ts
@@ -1,9 +1,11 @@
 import { useMutation, useQueryClient } from '@tanstack/react-query';
 
 import { addFestivalTimeTable } from '@shared/apis/time-table/festival-button';
-import { FESTIVAL_BUTTON_QUERY_KEY } from '@shared/apis/time-table/festival-button-queries';
 import { deleteFestivalTimetables } from '@shared/apis/time-table/festival-timetable';
-import { FESTIVAL_TIMETABLE_QUERY_KEY } from '@shared/apis/time-table/festival-timetable-queries';
+import {
+  FESTIVAL_BUTTON_QUERY_KEY,
+  FESTIVAL_TIMETABLE_QUERY_KEY,
+} from '@shared/constants/query-key';
 
 export const useDeleteTimeTableFestival = () => {
   const queryClient = useQueryClient();
