--- apps/client/src/shared/apis/confeti-detail/get-festival-to-add-queries.ts
@@ -1,20 +1,0 @@
-import { infiniteQueryOptions } from '@tanstack/react-query';
-
-import { GetFestivalToAddResponse } from '@shared/types/get-festival-to-add-response';
-
-import { getFestivalToAdd } from './get-festival-to-add';
-
-export const GET_FESTIVAL_TO_ADD_QUERY_KEY = {
-  ALL: ['getFestivalToAdd'],
-  LIST: () => [...GET_FESTIVAL_TO_ADD_QUERY_KEY.ALL, 'list'],
-} as const;
-
-export const GET_FESTIVAL_TO_ADD_QUERY_OPTIONS = {
-  LIST: () =>
-    infiniteQueryOptions<GetFestivalToAddResponse, Error>({
-      queryKey: GET_FESTIVAL_TO_ADD_QUERY_KEY.LIST(),
-      queryFn: ({ pageParam }) => getFestivalToAdd(pageParam as number),
-      initialPageParam: undefined,
-      getNextPageParam: (lastPage) => lastPage.nextCursor || undefined,
-    }),
-};
