+++ apps/client/src/shared/apis/time-table/festival-button-queries.ts
@@ -1,10 +1,8 @@
 import { queryOptions } from '@tanstack/react-query';
 
-import { getFestivalButton } from './festival-button';
+import { FESTIVAL_BUTTON_QUERY_KEY } from '@shared/constants/query-key';
 
-export const FESTIVAL_BUTTON_QUERY_KEY = {
-  ALL: ['festivalButton'],
-} as const;
+import { getFestivalButton } from './festival-button';
 
 export const FESTIVAL_BUTTON_QUERY_OPTIONS = {
   FESTIVAL_BUTTON: () =>
