+++ apps/client/src/shared/apis/time-table/festival-timetable-queries.ts
@@ -1,34 +1,23 @@
 import { queryOptions } from '@tanstack/react-query';
 
 import {
+  FESTIVAL_TIMETABLE_QUERY_KEY,
+  TIME_TABLE_CREATION_HISTORY_QUERY_KEY,
+} from '@shared/constants/query-key';
+
+import {
   getFestivalTimetable,
   getTimeTableCreationHistory,
 } from './festival-timetable';
 
-export const FESTIVAL_TIMETABLE_QUERY_KEY = {
-  ALL: ['festivalTimetable'],
-  DELETE_TIME_TABLE_FESTIVAL: (festivalId: number) => [
-    ...FESTIVAL_TIMETABLE_QUERY_KEY.ALL,
-    festivalId,
-  ],
-} as const;
-
 export const FESTIVAL_TIMETABLE_QUERY_OPTIONS = {
-  FESTIVAL_TIMETABLE: (festivalDateId: number) => {
-    const options = queryOptions({
-      queryKey: [...FESTIVAL_TIMETABLE_QUERY_KEY.ALL, festivalDateId],
-      queryFn: () => getFestivalTimetable(festivalDateId),
-    });
-    return {
-      ...options,
-      enabled: !!festivalDateId,
-    };
-  },
+  FESTIVAL_TIMETABLE: (festivalId: number) =>
+    queryOptions({
+      queryKey:
+        FESTIVAL_TIMETABLE_QUERY_KEY.DELETE_TIME_TABLE_FESTIVAL(festivalId),
+      queryFn: () => getFestivalTimetable(festivalId),
+    }),
 };
-
-export const TIME_TABLE_CREATION_HISTORY_QUERY_KEY = {
-  ALL: ['timeTableCreationHistory'],
-} as const;
 
 export const TIMETABLE_ONBOARDING_QUERY_KEY = {
   TIME_TABLE_CREATION_HISTORY: () =>
